+++ rados/rados_test.go
@@ -1200,16 +1200,3 @@ func (suite *RadosTestSuite) TestOpenIOContextInvalidPool() {
 func TestRadosTestSuite(t *testing.T) {
 	suite.Run(t, new(RadosTestSuite))
 }
-
-func TestRadosError(t *testing.T) {
-	err := getRadosError(0)
-	assert.NoError(t, err)
-
-	err = getRadosError(-5) // IO error
-	assert.Error(t, err)
-	assert.Equal(t, err.Error(), "rados: ret=5, Input/output error")
-
-	err = getRadosError(345) // no such errno
-	assert.Error(t, err)
-	assert.Equal(t, err.Error(), "rados: ret=345")
-}
